refactor: CQRS 패턴 적용을 위한 TrackingService 클래스 분리
- TrackingCommandService 구현: 산책 기록의 생성, 수정, 삭제(CUD) 등 데이터 변경 상태를 다루는 비즈니스 로직 전담

- TrackingQueryService 구현: 데이터 변경이 없는 산책 기록 조회(Read) 로직 전담 및 읽기 전용 트랜잭션 적용으로 성능 최적화

- TrackingVerification 구현: 여러 서비스에서 공통으로 사용되는 펫/산책 기록의 소유권 검증 로직을 별도 컴포넌트로 분리

- TrackingController 의존성 수정: 엔드포인트의 목적(Command/Query)에 맞게 분리된 서비스를 주입받아 사용하도록 개선
