+++ src/main/java/org/zerock/puppyrun/tracking/service/TrackingCommandService.java
@@ -7,15 +7,12 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
-import org.zerock.puppyrun.common.exception.ResourceNotFoundException;
-import org.zerock.puppyrun.common.exception.UserForbiddenException;
 import org.zerock.puppyrun.diary.entity.Diary;
 import org.zerock.puppyrun.diary.repository.DiaryRepository;
 import org.zerock.puppyrun.member.entity.Member;
 import org.zerock.puppyrun.member.repository.MemberRepository;
 import org.zerock.puppyrun.tracking.controller.request.ChangeVisibilityRequest;
 import org.zerock.puppyrun.tracking.controller.request.RegisterTrackingRequest;
-import org.zerock.puppyrun.tracking.controller.response.MainTrackingResponse;
 import org.zerock.puppyrun.tracking.controller.request.UpdateTrackingRequest;
 import org.zerock.puppyrun.tracking.DTO.UpdateTrackingDTO;
 import org.zerock.puppyrun.tracking.controller.response.TrackingDetailResponse;
@@ -27,30 +24,16 @@ import org.zerock.puppyrun.tracking.repository.TrackingRepository;
 @Service
 @RequiredArgsConstructor
 @Slf4j
-@Transactional(readOnly = true)
-public class TrackingService {
+@Transactional
+public class TrackingCommandService {
     private final DiaryRepository diaryRepository;
     private final TrackingRepository trackingRepository;
     private final MemberRepository memberRepository;
-
-    /**
-     * 산책 상세 조회
-     */
-    public Tracking findTrackingWithOwnershipCheck(UUID memberId, UUID trackingId) {
-        Tracking tracking = trackingRepository.findById(trackingId)
-                .orElseThrow(() -> new ResourceNotFoundException("해당 산책 기록을 찾을 수 없습니다."));
-
-        if (tracking.isNotOwner(memberId)) {
-            throw new UserForbiddenException("해당 산책 기록에 대한 권한이 없습니다.");
-        }
-
-        return tracking;
-    }
+    private final TrackingVerification trackingVerification;
 
     /**
      * 산책 저장
      */
-    @Transactional
     public void saveTracking(UUID memberId, RegisterTrackingRequest request, List<MultipartFile> images) {
         Member member = memberRepository.findByIdOrThrow(memberId);
 
@@ -73,15 +56,15 @@ public class TrackingService {
                 .path(path)
                 .build();
 
-        trackingRepository.save(tracking);
+        Tracking savedTracking = trackingRepository.save(tracking);
+
     }
 
     /**
      * 산책 정보 수정
      */
-    @Transactional
     public TrackingDetailResponse updateTracking(UUID memberId, UUID trackingId, UpdateTrackingRequest request) {
-        Tracking tracking = findTrackingWithOwnershipCheck(memberId, trackingId);
+        Tracking tracking = trackingVerification.ownershipCheck(memberId, trackingId);
 
         UpdateTrackingDTO updateTrackingDTO = UpdateTrackingDTO.builder()
                 .endedAt(request.endedAt())
@@ -101,9 +84,8 @@ public class TrackingService {
     /**
      * 산책 공개 여부 변경 (Lightweight Update)
      */
-    @Transactional
     public void changeTrackingVisibility(UUID memberId, UUID trackingId, ChangeVisibilityRequest request) {
-        Tracking tracking = findTrackingWithOwnershipCheck(memberId, trackingId);
+        Tracking tracking = trackingVerification.ownershipCheck(memberId, trackingId);
         Visibility visibility = Visibility.from(request.visibility());
         // 공개 여부 변경
         tracking.changeVisibility(visibility);
@@ -112,33 +94,12 @@ public class TrackingService {
     /**
      * 산책 기록 삭제
      */
-    @Transactional
     public void deleteTracking(UUID memberId, UUID trackingId) {
-        Tracking tracking = findTrackingWithOwnershipCheck(memberId, trackingId);
+        Tracking tracking = trackingVerification.ownershipCheck(memberId, trackingId);
 
         // 연관된 일기가 있다면 tracking_id를 null로 변경
         diaryRepository.findByTrackingId(trackingId)
                 .ifPresent(Diary::unsetTracking);
         trackingRepository.delete(tracking);
-    }
-
-    /**
-     * 산책 리스트 조회
-     */
-    public MainTrackingResponse getTrackingListResponse(UUID memberId) {
-        List<Tracking> trackingList = trackingRepository.findAllByMemberId(memberId);
-        return MainTrackingResponse.from(trackingList);
-    }
-
-    /**
-     * 산책 상세 조회
-     */
-    public TrackingDetailResponse getTrackingResponse(UUID memberId, UUID trackingId) {
-        Tracking tracking = findTrackingWithOwnershipCheck(memberId, trackingId);
-
-        UUID diaryId = diaryRepository.findIdByTrackingId(trackingId)
-                .orElse(null);
-
-        return TrackingDetailResponse.of(tracking, diaryId);
     }
 }
